rgph-dataviz/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
